+++ src/test/groovy/ru/vyarus/gradle/plugin/python/cmd/PythonCliTest.groovy
@@ -125,5 +125,28 @@ class PythonCliTest extends AbstractTest {
         python.clearExtraArgs()
         then: 'set'
         python.extraArgs.empty
+
+        when: "set vars"
+        python.environment('foo', 1)
+        python.environment('bar', 2)
+        then: 'set'
+        python.envVars == ['foo': 1, 'bar' : 2]
+
+        when: 'clear vars'
+        python.clearEnvironment()
+        then: 'no vars'
+        python.envVars.size() == 0
+
+        when: 'mass variables set'
+        python.environment(['foo': 2, 'bar' : 3])
+        python.environment(['foo': 1, 'baz' : 4])
+        then: 'aggregated'
+        python.envVars == ['foo': 1, 'bar' : 3, 'baz': 4]
+
+        when: 'additional var'
+        python.environment('sample': 'sam')
+        then: 'aggregated'
+        python.envVars == ['foo': 1, 'bar' : 3, 'baz': 4, 'sample': 'sam']
+
     }
 }